requirements.txt
pycryptodome==3.11.0
pycryptodomex==3.11.0